+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [Changed] `--completion` option to `--gen-completion`
 * [Changed] `--completion-out` option to `--gen-completion-out`
 * [Added] `--load-config` option to specify config file [#394](https://github.com/dalance/procs/issues/394)
+* [Added] `--use-config` option to specify built-in config [#152](https://github.com/dalance/procs/pull/152)
 * [Added] `show_header` and `show_footer` config [#405](https://github.com/dalance/procs/issues/405)
 * [Fixed] hang on terminals which ignore DSR request [#288](https://github.com/dalance/procs/issues/288)